requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.0
boto3==1.42.6
botocore==1.42.6
captum==0.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
filelock==3.20.0
filetype==1.2.0
fire==0.7.1
fonttools==4.61.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpickle==4.1.1
kagglehub==0.3.13
kiwisolver==1.4.9
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
megadetector==5.0.28
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.12.0
onnx2torch==1.5.15
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
pandas==2.3.3
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
protobuf==3.20.1
pybboxes==0.1.6
pycocotools==2.0.10
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
Pyqtree==1.0.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
reverse_geocoder==1.5.1
roboflow==1.2.11
s3transfer==0.16.0
sahi==0.11.36
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
smmap==5.0.2
speciesnet==5.0.0
stringzilla==4.4.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics-yolov5==0.1.1
urllib3==2.6.1
Werkzeug==3.1.4
yolov5==7.0.11
yolov9pip==0.0.4